Make local dib file cleanup method a static method
By making this a static method, we can allow code other than the
CleanupWorker thread to call it for DIB file removal.

This change helped to identify that the builder unit test
test_cleanup_failed_image_build was not valid because the fake
dib creation process was exiting before creating any files.
Since the test was attempting to validate those files were
supposed to be deleted, it would always succeed. Moving the
exit to after file creation fixes that.

Change-Id: Iaa5f02abbaebd9866b6cc16c977e0d42d804eee1